🚀add Login & MyPage

+++ src/main/kotlin/kr/ac/kumoh/s20190645/rakuten/controller/PlayerController.kt
@@ -3,7 +3,8 @@ package kr.ac.kumoh.s20190645.rakuten.controller
 import kr.ac.kumoh.s20190645.rakuten.service.PlayerService
 import org.slf4j.LoggerFactory
 import org.springframework.http.ResponseEntity
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder
+import org.springframework.security.access.prepost.PreAuthorize
+import org.springframework.security.core.Authentication
 import org.springframework.stereotype.Controller
 import org.springframework.ui.Model
 import org.springframework.web.bind.annotation.*
@@ -17,56 +18,56 @@ class PlayerController(
 
     @GetMapping("/")
     fun home(): String {
-        return "index"
+        return "/Normal/index"
     }
 
     @GetMapping("/list")
     fun list(model: Model): String {
         val result = playerService.findAllPlayer()
         model.addAttribute("players", result)
-        return "list"
+        return "/Normal/list"
     }
 
     @GetMapping("/random")
     fun random(model: Model): String {
         val result = playerService.findRandom()
         model.addAttribute("player", result)
-        return "random"
+        return "/Normal/random"
     }
 
     @GetMapping("/info/{backNumber}")
-    fun info(model: Model, @PathVariable backNumber: Int): Any {
-        return try {
-            val result = playerService.findPlayer(backNumber)
-            model.addAttribute("player", result)
-            "info"
-        } catch (e: Exception) {
-            logger.error("error!", e)
-            return ResponseEntity.status(400).body("登録された選手がおりません")
-        }
+    fun info(model: Model, @PathVariable backNumber: Int): String {
+        val result = playerService.findPlayer(backNumber)
+        model.addAttribute("player", result)
+        return "/Normal/info"
     }
 
+    @PreAuthorize("hasRole('ADMIN')")
     @GetMapping("/add")
     fun add(): String {
-        return "add"
+        return "/Operation/add"
     }
 
+    @PreAuthorize("hasRole('ADMIN')")
     @PostMapping("/playerAdd")
     fun addPlayer(model: Model, @RequestParam params: Map<String, String>): String {
         val name = params["name"] ?: "unknown"
         val backNumber = params["backNumber"]?.toIntOrNull()
         val addResult = playerService.addPlayer(name, backNumber)
         model.addAttribute("result", addResult)
-        return "addResult"
+        return "/Operation/addResult"
     }
 
+    
+    @PreAuthorize("hasRole('ADMIN')")
     @GetMapping("/update/{backNumber}")
     fun updateForm(model: Model, @PathVariable backNumber: Int): String {
         val found = playerService.findPlayer(backNumber)
         model.addAttribute("player", found)
-        return "update"
+        return "/Operation/update"
     }
 
+    @PreAuthorize("hasRole('ADMIN')")
     @PostMapping("/playerUpdate")
     fun update(model: Model, @RequestParam params: Map<String, String>): String {
         val info = params["info"]?.toIntOrNull()
@@ -75,18 +76,19 @@ class PlayerController(
         val backNumber = params["backNumber"]?.toIntOrNull()
         val result = playerService.updatePlayer(found, name, backNumber)
         model.addAttribute("result", result)
-        return "updateResult"
+        return "/Operation/updateResult"
     }
 
+    @PreAuthorize("hasRole('ADMIN')")
     @PostMapping("/delete")
     @ResponseBody
-    fun deletePlayer(@RequestBody params: Map<String, String>): String {
+    fun deletePlayer(@RequestBody params: Map<String, String>,auth: Authentication): String {
+        if (auth.authorities.none { it.authority == "ROLE_ADMIN" }) {
+            return "/access-denied"
+        }
         val number = params["playerNumber"]?.toIntOrNull()
         if (playerService.deletePlayer(number) != 1)
             return "/info/${number}"
         return "/list"
     }
-
-
-
 }
